Nishang

+++ app/Http/Controllers/Modules/StudentController.php
@@ -3,6 +3,7 @@
 
 namespace App\Http\Controllers\Modules;
 
+use App\AnnualClass;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
@@ -10,37 +11,18 @@ use Illuminate\Support\Facades\Hash;
 use DateTime;
 
 class StudentController extends Controller{
-
-    public function genMat($class, $year){
-        $class = \App\Classes::find($class);
-        $count = $class->students($year)->count();
-        return $class->abbreviations.$count.$this->getSection($class->id, $year);
-    }
-
-    public function getSection($class, $year){
-        $class = \App\Classes::find($class);
-        $sections = ['A','B','C','D','E','F'];
-        $count = $class->students($year)->count();
-        $position = ($count - ($count % $class->limit))/$class->limit;
-        if($position > 5){
-            return $sections[5];
-        }else{
-            return $sections[$position];
-        }
-    }
-
     public function index(Request $request){
         $data['students'] =[];
         $data['year'] = $request->year?$request->year:getYear();
         if(!$request->class){
             $data['students'] = \App\Student::orderBy('created_at','DESC')->get();
         }else{
-            $class = \App\ClassSection::find(\request('class'));
-            $data['students'] = $class->students($data['year']);
+            $class = \App\AnnualClass::find(\request('class'));
+            $data['students'] = $class->student;
         }
+        $data['class'] = \App\AnnualClass::find(\request('class'));
         return view('student.index')->with($data);
     }
-
     public function show(Request $request, $slug){
         $data['student'] = \App\Student::whereSlug($slug)->first();
         $data['year'] = $request->year?$request->year:getYear();
@@ -90,15 +72,6 @@ class StudentController extends Controller{
                 $student->phone = $request->phone;
                 $student->save();
 
-                $class = \App\Classes::find($student->class);
-                $classR = $student->classR($request->admission_year);
-                $classR->delete();
-                $studentClass = \App\StudentsClass::create([
-                    'student_id'=> $student->id,
-                    'class_id'=> $student->class,
-                    'year_id'=> $request->admission_year,
-                    'section_id'=> $this->getSection($request->class, getYear()),
-                ]);
                 \DB::commit();
                 $request->session()->flash('success', "Student updated successfully");
             }catch(\Exception $e){
@@ -111,8 +84,6 @@ class StudentController extends Controller{
         }
         return redirect()->to(route('student.index'));
     }
-
-
     public function store(Request $request)
     {
         if (\Auth::user()->can('create_student')) {
@@ -139,14 +110,12 @@ class StudentController extends Controller{
                 $input = $request->all();
                 $input['slug'] = str_replace("/","",$slug);
                 $input['photo'] = $image;
-                $input['matricule'] = $this->genMat($request->class, getYear());
+                $input['matricule'] = genMat($request->class, getYear());
                 $student = \App\Student::create($input);
-                $class = \App\Classes::find($student->class);
-                $studentClass = \App\StudentsClass::create([
+
+                 \App\StudentsClass::create([
                     'student_id'=> $student->id,
-                    'class_id'=> $student->class,
-                    'year_id'=> $request->admission_year,
-                    'section_id'=> $this->getSection($request->class, getYear()),
+                    'class_id'=> getSection($request->class, getYear())->id
                 ]);
                 \DB::commit();
                 $request->session()->flash('success', "Student Created successfully");
@@ -167,7 +136,7 @@ class StudentController extends Controller{
             if($student == null){
                 abort(404);
             }
-           if($student->feePayment->count() == 0 ){
+           if($student->feePayment->count() == 0 && $student->discount->count() == 0 && $student->hasMany('App\StudentsClass','student_id')->count() == 0 ){
                $student->delete();
                $request->session()->flash('success', "Student Deleted successfully");
            }else{
@@ -179,5 +148,127 @@ class StudentController extends Controller{
         }
 
         return redirect()->to(route('student.index'));
+    }
+
+    public function promote(Request $request){
+        $students = new \App\Student();
+        $year = getYear();
+        if($request->year){
+            $year = $request->year;
+        }
+
+        if($request->next_year){
+            if($request->next_year < $year){
+                $request->session()->flash('error', "Invalid promotion academic year");
+            }
+        }
+
+        if($request->class){
+            $students= \App\Classes::find($request->class)->students($year);
+        }
+
+        $data['students'] = $students;
+        return view('student.promote')->with($data);
+    }
+
+    public function promoteSubmit(Request $request){
+        $this->validate($request, [
+            'class' => 'required',
+            'students' => 'required',
+            'year' => 'required',
+            'next_year' => 'required',
+        ]);
+        try{
+            $class = \App\Classes::find($request->class);
+            $newClass = $class;
+            $students = $newClass->parent->students($request->next_year);
+           foreach ($request->students as $stud){
+               $student = \App\Student::find($stud);
+               if(!$this->isMember($student, $students)){
+                   $s =  \App\StudentsClass::create([
+                       'student_id'=> $student->id,
+                       'class_id'=> getSection($newClass->next_class, $request->next_year)->id
+                   ]);
+               }
+           }
+           $request->session()->flash('success', "Student Promoted Successfully");
+           return redirect()->back();
+        }catch(\Exception $e){
+           
+            $request->session()->flash('error', "Something went wrong");
+        }
+    }
+
+    public function changeClass(Request $request){
+        return view('student.search');
+    }
+
+    public function isMember( $student, $students){
+        $isMember = false;
+        foreach($students as $stud){
+            $isMember = $isMember || ($stud->student_id === $student->id);
+        }
+           
+        return $isMember;
+    }
+
+
+    public function changeClassForm(Request $request, $student){
+        $data['student'] = \App\Student::findOrFail($student);
+        return view('student.changeClass')->with($data);
+    }
+
+    public function changeClassFormPost(Request $request, $student){
+        $this->validate($request, [
+            'class' => 'required_without:remove',
+            'student' => 'required',
+            'current_class' => 'required',
+        ]);
+
+        $student = \App\Student::findOrFail($student);
+
+        try{
+            \DB::beginTransaction();
+            $classR = $student->classR($student->sClass()->id);
+
+            if($classR && $student->hasMany('App\StudentsClass','student_id')->count() > 1){
+                $classR->delete();
+            }elseif(isset($request->remove)){
+                $classR->delete();
+                if ($request->user()->can('delete_student')) {
+                    $student = \App\Student::whereSlug($id)->first();
+                    if($student == null){
+                        abort(404);
+                    }
+                   if($student->feePayment->count() == 0 && $student->discount->count() == 0 && $student->hasMany('App\StudentsClass','student_id')->count() == 0 ){
+                       $student->delete();
+                       $request->session()->flash('success', "Student Deleted successfully");
+                   }else{
+                       $request->session()->flash('error', "Cant Delete Student, has some transaction saved");
+                   }
+        
+                }else{
+                    $request->session()->flash('error', "Not allowed to perform this action");
+                }
+            }
+
+          if(!isset($request->remove)){
+            $studentClass = \App\StudentsClass::create([
+                'student_id'=> $student->id,
+                'class_id'=> getSection($request->class, getYear())->id
+            ]);
+          }
+
+
+
+            \DB::commit();
+            $request->session()->flash('success', "Student Migrated Successfully");
+        }catch(\Exception $e){
+            \DB::rollback();
+            //echo $e;
+        $request->session()->flash('error', "Something went wrong");
+       }
+
+         return redirect()->back();
     }
 }
